blog_api/serializers.py
from rest_framework import serializers
from blog.models import Post


class PostSerializer(serializers.ModelSerializer):
    class Meta:
        fields = ('id', 'title', 'user', 'excerpt', 'content', 'status')
        model = Post
